refactor: ProductAccordion 분리
